backend/requirements.txt
# API
fastapi
uvicorn[standard]

# Banco de dados
sqlalchemy
alembic
psycopg2-binary

# Validação e schemas
pydantic[email]

# Autenticação
passlib[bcrypt]
bcrypt

# Variáveis de ambiente
python-dotenv

# Processamento de imagens
pillow
numpy

# IA (TensorFlow)
tensorflow